feat(sauce like controller and route): add possibility to like/dislike a sauce
creation of a controller that allows an authenticated user to like/dislike a sauce. User can not
like a sauce several times without first removing its previous like. It is also impossible for him
to like AND dislike a sauce. Through an array of userId,  it is verified that the user hasn't
already like/dislike a sauce.

BREAKING CHANGE: application is now completely functionnal, but needs refactor and security fix for
images